feat(publish-helper): extract into helper
